Create ActiveContractsClient interface
