+++ src/client/DamlLedgerClient.ts
@@ -30,6 +30,7 @@ import {TimeServiceClient} from "../generated/com/digitalasset/ledger/api/v1/tes
 import {TransactionServiceClient} from "../generated/com/digitalasset/ledger/api/v1/transaction_service_grpc_pb";
 import {ResetServiceClient} from "../generated/com/digitalasset/ledger/api/v1/testing/reset_service_grpc_pb";
 import {NodeJsCommandClient} from "./NodeJsCommandClient";
+import {NodeJsActiveContractsClient} from "./NodeJsActiveContractsClient";
 
 /**
  * A {@link LedgerClient} implementation that connects to an existing Ledger and provides clients to query it. To use the {@link DamlLedgerClient}
@@ -56,7 +57,7 @@ export class DamlLedgerClient implements LedgerClient {
         reporter: ValidationReporter
     ) {
         this.ledgerId = ledgerId;
-        this._activeContractsClient = new ActiveContractsClient(
+        this._activeContractsClient = new NodeJsActiveContractsClient(
             ledgerId,
             new ActiveContractsServiceClient(address, credentials),
             reporter
